+++ HISTORY.rst
@@ -28,6 +28,7 @@ TODO
 
 * Update requirements
 * Import data from Hungary [#53](https://github.com/cnr-ibba/SMARTER-database/issues/53)
+* Create a new sample when having the same ``original_id`` in dataset but for a different breed
 
 0.4.4 (2022-02-28)
 ------------------